requirements.txt
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
requests==2.32.3
